requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
scikit-learn>=1.2.0
tqdm>=4.65.0
wandb>=0.15.0
PyYAML>=6.0
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
timm>=0.9.0
ftfy>=6.1.0
regex>=2023.0.0
